+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added the ability to show current tenant information.
 - Fixed a bug where the application icon wasn't reset after opening the manifest.
+- Added better error handling to detect user not being logged in.
 
 ## 0.2.0